ai-placement-platform/backend/requirements.txt
# ── Web Framework ─────────────────────────────
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# ── Database ──────────────────────────────────
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ── Cache ─────────────────────────────────────
redis==5.0.1
hiredis==2.3.2

# ── Auth & Security ───────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# ── AI / ML ───────────────────────────────────
openai==1.12.0
spacy==3.7.4
sentence-transformers==2.4.0
numpy==1.26.4
scikit-learn==1.4.0

# ── Data Validation ───────────────────────────
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# ── HTTP Client ───────────────────────────────
httpx==0.26.0
aiohttp==3.9.3

# ── PDF / Document Parsing ────────────────────
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3

# ── Google APIs ───────────────────────────────
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.119.0

# ── Email ─────────────────────────────────────
aiosmtplib==3.0.1
jinja2==3.1.3

# ── Background Tasks ──────────────────────────
celery==5.3.6
apscheduler==3.10.4

# ── Testing ───────────────────────────────────
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
factory-boy==3.3.0

# ── Utilities ─────────────────────────────────
python-dateutil==2.8.2
pytz==2024.1
phonenumbers==8.13.30
pillow==10.2.0
